+++ persistent-postgresql/test/main.hs
@@ -60,7 +60,6 @@ import qualified TreeTest
 import qualified UniqueTest
 import qualified UpsertTest
 import qualified UpsertWhere
-import qualified PgIntervalTest
 
 type Tuple = (,)