+++ persistent-postgresql/ChangeLog.md
@@ -1,5 +1,17 @@
 # Changelog for persistent-postgresql
 
+## 2.13.2.1
+
+* [#1331](https://github.com/yesodweb/persistent/pull/1331)
+    * Fixes a bug where `upsertWhere` would fail on a database table with
+      `MigrationOnly` fields.
+
+## 2.13.2.0
+* [#1316](https://github.com/yesodweb/persistent/pull/1316)
+  * Expose some internals in the new `Database.Persist.Postgresql.Internal` module.
+    This gives access to the `P` newtype, which is used for de-serializing `PersistValue`s
+    from `postgresql-simple` code.
+
 ## 2.13.1.0
 
 * [#1305](https://github.com/yesodweb/persistent/pull/1305)